refactor: move shared tree ui to shared
