+++ src/pages/dead-code/widgets/DeadCodeExplorer/ui/DeadCodeCategory.tsx
@@ -7,7 +7,6 @@ import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import { useMemo } from 'react';
 import { activeActivityPageIdAtom } from '@/app/model/activityPageAtoms';
-import { FileTreeItem } from '@/components/ide/FileTreeItem.tsx';
 import { filesAtom, viewModeAtom } from '@/entities/AppView/model/atoms';
 import { FileIcon } from '@/entities/SourceFileNode/ui/FileIcon.tsx';
 import { buildDeadCodeTree } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/buildDeadCodeTree.ts';
@@ -23,6 +22,7 @@ import { CategoryCheckbox } from '@/features/Code/CodeAnalyzer/DeadCodeSelection
 import { targetLineAtom } from '@/features/File/Navigation/model/atoms.ts';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile.ts';
 import { Checkbox } from '@/shared/ui/Checkbox';
+import { FileTreeItem } from '@/shared/ui/FileTreeItem';
 import { TreeView } from '@/shared/ui/TreeView/TreeView.tsx';
 import { DeadCodeFolderItem } from './DeadCodeFolderItem.tsx';
