+++ src/pages/dead-code/widgets/DeadCodePanel/DeadCodePanel.tsx
@@ -6,7 +6,6 @@
 import { useAtomValue } from 'jotai';
 import { Sparkles } from 'lucide-react';
 import { useRef, useState } from 'react';
-import { Sidebar } from '@/components/ide/Sidebar.tsx';
 import { useDeadCodeAnalysis } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/useDeadCodeAnalysis.ts';
 import {
   deadCodeResultsAtom,
@@ -14,6 +13,7 @@ import {
 } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/model/atoms.ts';
 import { RefactoringPromptDialog } from '@/features/RefactoringPrompt/RefactoringPromptDialog.tsx';
 import { Button } from '@/shared/ui/Button';
+import { Sidebar } from '@/shared/ui/Sidebar';
 import IDEScrollView from '@/widgets/MainContents/IDEScrollView/IDEScrollView.tsx';
 import { DeadCodeExplorer } from '../DeadCodeExplorer/DeadCodeExplorer.tsx';
 import { DeadCodePanelHeader } from './ui/DeadCodePanelHeader.tsx';
