+++ src/shared/ui/FileTreeItem.tsx
