+++ src/widgets/FileExplorer/FileExplorer.tsx
@@ -7,10 +7,10 @@ import { useAtom, useAtomValue } from 'jotai';
 import { Folder, FolderOpen } from 'lucide-react';
 import type React from 'react';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { FileTreeItem } from '@/components/ide/FileTreeItem';
 import { filesAtom, focusedFolderAtom } from '@/entities/AppView/model/atoms';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile';
 import { activeTabAtom, openedTabsAtom } from '@/features/File/OpenFiles/model/atoms';
+import { FileTreeItem } from '@/shared/ui/FileTreeItem';
 import { FileIcon } from '../../entities/SourceFileNode/ui/FileIcon.tsx';
 import { useTreeKeyboardNavigation } from '../../shared/hooks/useTreeKeyboardNavigation';
 import { TreeView } from '../../shared/ui/TreeView/TreeView';
